agentv1/requirements.txt
langchain-google-genai 
geopy 
requests
langchain
langgraph
IPython
python-dotenv
langchain_chroma
langchain-google-genai
google-generativeai
langchain-chroma
rank-bm25
pdfplumber
python-pptx
python-docx
requests
flask 
flask_cors 
dotenv
google-api-python-client
google-auth
google-auth-oauthlib
